requirements.txt
appdirs==1.4.4
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5
certifi==2020.4.5.1
chardet==4.0.0
colorama==0.4.3
cvlib==0.2.6
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.3
django-crispy-forms==1.9.2
django-debug-toolbar==3.1.1
django-filter==2.3.0
django-heroku==0.3.1
easy-thumbnails==2.7
face-recognition-models==0.3.0
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
imageio==2.9.0
imutils==0.5.4
isort==4.3.21
joblib==1.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
matplotlib==3.3.3
mccabe==0.6.1
mysql-connector-python==8.0.26
networkx==2.5
numpy==1.19.5
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
pandas==1.2.1
Pillow==8.4.0
pipenv==2021.11.15
progressbar==2.5
protobuf==3.18.1
psycopg2==2.9.2
pycodestyle==2.5.0
pylint==2.4.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-mnist==0.7
pytz==2019.3
PyWavelets==1.1.1
requests==2.25.1
scikit-image==0.18.1
scikit-learn==0.24.0
scipy==1.6.0
six==1.14.0
sqlparse==0.3.1
threadpoolctl==2.1.0
tifffile==2021.1.14
toml==0.10.2
urllib3==1.26.2
virtualenv==20.4.7
virtualenv-clone==0.5.7
whitenoise==5.3.0
wincertstore==0.2
wrapt==1.11.2
